FIRE-34747: Move FSJointPose to own class
