+++ indra/newview/fsposingmotion.cpp
@@ -48,7 +48,7 @@ LLMotion::LLMotionInitStatus FSPosingMotion::onInitialize(LLCharacter *character
         if (!targetJoint)
             continue;
 
-        FSJointPose jointPose = FSJointPose(targetJoint);
+        FSJointPose jointPose = FSJointPose(targetJoint, POSER_JOINT_STATE);
         mJointPoses.push_back(jointPose);
 
         addJointState(jointPose.getJointState());
@@ -213,7 +213,7 @@ void FSPosingMotion::setJointState(LLJoint* joint, U32 state)
     addJointState(jointPose->getJointState());
 }
 
-FSPosingMotion::FSJointPose* FSPosingMotion::getJointPoseByJointName(const std::string& name)
+FSJointPose* FSPosingMotion::getJointPoseByJointName(const std::string& name)
 {
     if (mJointPoses.size() < 1)
         return nullptr;
@@ -285,251 +285,3 @@ constexpr size_t MaximumUndoQueueLength = 20;
 /// The constant time interval, in seconds, specifying whether an 'undo' value should be added.
 /// </summary>
 constexpr std::chrono::duration<double> UndoUpdateInterval = std::chrono::duration<double>(0.3);
-
-
-void FSPosingMotion::FSJointPose::addLastPositionToUndo()
-{
-    if (mUndonePositionIndex > 0)
-    {
-        for (int i = 0; i < mUndonePositionIndex; i++)
-            mLastSetPositions.pop_front();
-
-        mUndonePositionIndex = 0;
-    }
-
-    mLastSetPositions.push_front(mTargetPosition);
-
-    while (mLastSetPositions.size() > MaximumUndoQueueLength)
-        mLastSetPositions.pop_back();
-}
-
-void FSPosingMotion::FSJointPose::addLastRotationToUndo()
-{
-    if (mUndoneRotationIndex > 0)
-    {
-        for (int i = 0; i < mUndoneRotationIndex; i++)
-            mLastSetRotations.pop_front();
-
-        mUndoneRotationIndex = 0;
-    }
-
-    mLastSetRotations.push_front(mTargetRotation);
-
-    while (mLastSetRotations.size() > MaximumUndoQueueLength)
-        mLastSetRotations.pop_back();
-}
-
-void FSPosingMotion::FSJointPose::addLastScaleToUndo()
-{
-    if (mUndoneScaleIndex > 0)
-    {
-        for (int i = 0; i < mUndoneScaleIndex; i++)
-            mLastSetScales.pop_front();
-
-        mUndoneScaleIndex = 0;
-    }
-
-    mLastSetScales.push_front(mTargetScale);
-
-    while (mLastSetScales.size() > MaximumUndoQueueLength)
-        mLastSetScales.pop_back();
-}
-
-LLVector3 FSPosingMotion::FSJointPose::getCurrentPosition()
-{
-    LLVector3 vec3;
-    LLJoint* joint = mJointState->getJoint();
-    if (!joint)
-        return vec3;
-
-    vec3 = joint->getPosition();
-    return vec3;
-}
-
-void FSPosingMotion::FSJointPose::setTargetPosition(const LLVector3& pos)
-{
-    auto timeIntervalSinceLastRotationChange = std::chrono::system_clock::now() - mTimeLastUpdatedPosition;
-    if (timeIntervalSinceLastRotationChange > UndoUpdateInterval)
-        addLastPositionToUndo();
-
-    mTimeLastUpdatedPosition = std::chrono::system_clock::now();
-    mTargetPosition.set(pos);
-}
-
-LLQuaternion FSPosingMotion::FSJointPose::getCurrentRotation()
-{
-    LLQuaternion quat;
-    LLJoint* joint = mJointState->getJoint();
-    if (!joint)
-        return quat;
-
-    quat = joint->getRotation();
-    return quat;
-}
-
-void FSPosingMotion::FSJointPose::setTargetRotation(const LLQuaternion& rot)
-{
-    auto timeIntervalSinceLastRotationChange = std::chrono::system_clock::now() - mTimeLastUpdatedRotation;
-    if (timeIntervalSinceLastRotationChange > UndoUpdateInterval)
-        addLastRotationToUndo();
-
-    mTimeLastUpdatedRotation = std::chrono::system_clock::now();
-    mTargetRotation.set(rot);
-}
-
-void FSPosingMotion::FSJointPose::applyDeltaRotation(const LLQuaternion& rot)
-{
-    auto timeIntervalSinceLastRotationChange = std::chrono::system_clock::now() - mTimeLastUpdatedRotation;
-    if (timeIntervalSinceLastRotationChange > UndoUpdateInterval)
-        addLastRotationToUndo();
-
-    mTimeLastUpdatedRotation = std::chrono::system_clock::now();
-    mTargetRotation = mTargetRotation * rot;
-}
-
-LLVector3 FSPosingMotion::FSJointPose::getCurrentScale()
-{
-    LLVector3 vec3;
-    LLJoint* joint = mJointState->getJoint();
-    if (!joint)
-        return vec3;
-
-    vec3 = joint->getScale();
-    return vec3;
-}
-
-void FSPosingMotion::FSJointPose::setTargetScale(LLVector3 scale)
-{
-    auto timeIntervalSinceLastScaleChange = std::chrono::system_clock::now() - mTimeLastUpdatedScale;
-    if (timeIntervalSinceLastScaleChange > UndoUpdateInterval)
-        addLastScaleToUndo();
-
-    mTimeLastUpdatedScale = std::chrono::system_clock::now();
-    mTargetScale.set(scale);
-}
-
-void FSPosingMotion::FSJointPose::undoLastPositionSet()
-{
-    if (mLastSetPositions.empty())
-        return;
-
-    if (mUndonePositionIndex == 0)  // at the top of the queue add the current
-        addLastPositionToUndo();
-
-    mUndonePositionIndex++;
-    mUndonePositionIndex = llclamp(mUndonePositionIndex, 0, mLastSetPositions.size() - 1);
-    mTargetPosition.set(mLastSetPositions[mUndonePositionIndex]);
-}
-
-void FSPosingMotion::FSJointPose::redoLastPositionSet()
-{
-    if (mLastSetPositions.empty())
-        return;
-
-    mUndonePositionIndex--;
-    mUndonePositionIndex = llclamp(mUndonePositionIndex, 0, mLastSetPositions.size() - 1);
-
-    mTargetPosition.set(mLastSetPositions[mUndonePositionIndex]);
-    if (mUndonePositionIndex == 0)
-        mLastSetPositions.pop_front();
-}
-
-void FSPosingMotion::FSJointPose::undoLastRotationSet()
-{
-    if (mLastSetRotations.empty())
-        return;
-
-    if (mUndoneRotationIndex == 0) // at the top of the queue add the current
-        addLastRotationToUndo();
-
-    mUndoneRotationIndex++;
-    mUndoneRotationIndex = llclamp(mUndoneRotationIndex, 0, mLastSetRotations.size() - 1);
-    mTargetRotation.set(mLastSetRotations[mUndoneRotationIndex]);
-}
-
-void FSPosingMotion::FSJointPose::redoLastRotationSet()
-{
-    if (mLastSetRotations.empty())
-        return;
-
-    mUndoneRotationIndex--;
-    mUndoneRotationIndex = llclamp(mUndoneRotationIndex, 0, mLastSetRotations.size() - 1);
-
-    mTargetRotation.set(mLastSetRotations[mUndoneRotationIndex]);
-    if (mUndoneRotationIndex == 0)
-        mLastSetRotations.pop_front();
-}
-
-void FSPosingMotion::FSJointPose::undoLastScaleSet()
-{
-    if (mLastSetScales.empty())
-        return;
-
-    if (mUndoneScaleIndex == 0)
-        addLastScaleToUndo();
-
-    mUndoneScaleIndex++;
-    mUndoneScaleIndex = llclamp(mUndoneScaleIndex, 0, mLastSetScales.size() - 1);
-    mTargetScale.set(mLastSetScales[mUndoneScaleIndex]);
-}
-
-void FSPosingMotion::FSJointPose::redoLastScaleSet()
-{
-    if (mLastSetScales.empty())
-        return;
-
-    mUndoneScaleIndex--;
-    mUndoneScaleIndex = llclamp(mUndoneScaleIndex, 0, mLastSetScales.size() - 1);
-
-    mTargetScale.set(mLastSetScales[mUndoneScaleIndex]);
-    if (mUndoneScaleIndex == 0)
-        mLastSetScales.pop_front();
-}
-
-void FSPosingMotion::FSJointPose::revertJointScale()
-{
-    LLJoint* joint = mJointState->getJoint();
-    if (!joint)
-        return;
-
-    joint->setScale(mBeginningScale);
-}
-
-void FSPosingMotion::FSJointPose::revertJointPosition()
-{
-    LLJoint* joint = mJointState->getJoint();
-    if (!joint)
-        return;
-
-    joint->setPosition(mBeginningPosition);
-}
-
-void FSPosingMotion::FSJointPose::revertCollisionVolume()
-{
-    if (!mIsCollisionVolume)
-        return;
-
-    LLJoint* joint = mJointState->getJoint();
-    if (!joint)
-        return;
-
-    joint->setRotation(mBeginningRotation);
-    joint->setPosition(mBeginningPosition);
-    joint->setScale(mBeginningScale);
-}
-
-void FSPosingMotion::FSJointPose::setJointStartRotations(LLQuaternion quat) { mBeginningRotation = mTargetRotation = quat; }
-
-FSPosingMotion::FSJointPose::FSJointPose(LLJoint* joint, bool isCollisionVolume)
-{
-    mJointState = new LLJointState;
-    mJointState->setJoint(joint);
-    mJointState->setUsage(POSER_JOINT_STATE);
-
-    mJointName = joint->getName();
-    mIsCollisionVolume = isCollisionVolume;
-
-    mBeginningRotation = mTargetRotation = joint->getRotation();
-    mBeginningPosition = mTargetPosition = joint->getPosition();
-    mBeginningScale = mTargetScale = joint->getScale();
-}
